#33 #35 add info api
